first pass move lookup to clickhosue
